requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
aiohttp==3.7.4.post0
    # via twitchio
altgraph==0.17.3
    # via pyinstaller
anyio==3.7.0
    # via httpcore
async-timeout==3.0.1
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
chardet==4.0.0
    # via aiohttp
charset-normalizer==3.1.0
    # via requests
cryptography==40.0.2
    # via -r requirements.in
datetime==5.1
    # via -r requirements.in
distlib==0.3.6
    # via virtualenv
filelock==3.12.0
    # via virtualenv
h11==0.14.0
    # via httpcore
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via -r requirements.in
identify==2.5.24
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iso8601==1.1.0
    # via twitchio
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nodeenv==1.8.0
    # via pre-commit
platformdirs==3.5.1
    # via virtualenv
pre-commit==3.3.2
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyinstaller==5.12.0
    # via -r requirements.in
pyinstaller-hooks-contrib==2023.3
    # via pyinstaller
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via datetime
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
twitchio==2.1.5
    # via -r requirements.in
typing-extensions==4.6.2
    # via
    #   aiohttp
    #   twitchio
urllib3==2.0.2
    # via requests
virtualenv==20.23.0
    # via pre-commit
yarl==1.9.2
    # via aiohttp
zope-interface==6.0
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
